Split traits to manage TOML dependency.
